# contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/reps:
    get:
      description: "Requires hrm:view. Filters: status, fieldManagerId, search.
        payment_details redacted unless hrm:edit."
      operationId: RepsController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            default: active
            type: string
            enum:
              - active
              - terminated
              - all
        - name: fieldManagerId
          required: false
          in: query
          description: Filter by field manager (user id).
          schema:
            type: string
        - name: search
          required: false
          in: query
          description: Search by name or rep_code (case-insensitive).
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: &a17
        - bearer: []
      summary: List reps
      tags: &a18
        - HRM / Reps
    post:
      description: Requires hrm:create. rep_code never reused → 409.
      operationId: RepsController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepDto"
      responses:
        "201":
          description: ""
      security: *a17
      summary: Create a rep
      tags: *a18
  /v1/reps/{id}:
    get:
      description: Requires hrm:view. payment_details redacted unless hrm:edit.
      operationId: RepsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a17
      summary: Get a rep
      tags: *a18
    patch:
      description: Requires hrm:edit. status=terminated requires termination_date.
        rep_code immutable.
      operationId: RepsController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRepDto"
      responses:
        "200":
          description: ""
      security: *a17
      summary: Edit / set field manager / terminate a rep
      tags: *a18
  /v1/reps/{id}/documents:
    get:
      description: Requires hrm:view. file_url shown only with hrm:edit.
      operationId: RepsController_listDocuments_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a17
      summary: List a rep's documents
      tags: *a18
    post:
      description: Requires hrm:edit. Stores a storage reference.
      operationId: RepsController_createDocument_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepDocumentDto"
      responses:
        "201":
          description: ""
      security: *a17
      summary: Attach a document to a rep
      tags: *a18
  /v1/reps/{id}/equipment:
    get:
      description: Requires hrm:view.
      operationId: RepsController_listEquipment_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a17
      summary: List a rep's equipment
      tags: *a18
    post:
      description: Requires hrm:edit. Deposit is exact Decimal.
      operationId: RepsController_assignEquipment_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepEquipmentDto"
      responses:
        "201":
          description: ""
      security: *a17
      summary: Assign equipment to a rep
      tags: *a18
...
  /v1/imports:
    post:
      description: Requires import:create. Stages + classifies rows (file upload
        stubbed — rows fed in the body). Nothing is written to live tables until
        commit.
      operationId: ImportController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateImportDto"
      responses:
        "201":
          description: ""
      security: &a45
        - bearer: []
      summary: Create a staged import batch
      tags: &a46
        - Data Import & Integration
    get:
      description: Requires import:view.
      operationId: ImportController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - staged
              - committed
              - failed
              - cancelled
        - name: source_type
          required: false
          in: query
          schema:
            type: string
            enum:
              - client_report
              - master_migration
              - balance_migration
        - name: import_type
          required: false
          in: query
          schema:
            type: string
            enum:
              - reps
              - clients
              - products
              - sales
              - holdback
              - clawback
              - mixed
      responses:
        "200":
          description: ""
      security: *a45
      summary: Import/migration history
      tags: *a46
  /v1/imports/{id}:
    get:
      description: Requires import:view. Rows with match status, counts, and reconcile
        total.
      operationId: ImportController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a45
      summary: Preview a staged batch
      tags: *a46
  /v1/imports/{id}/reconcile:
    post:
      description: "Requires import:edit. Resolutions: match / edit / ignore.
        Recomputes counts."
      operationId: ImportController_reconcile_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconcileDto"
      responses:
        "201":
          description: ""
      security: *a45
      summary: Resolve unmatched/ambiguous rows
      tags: *a46
  /v1/imports/{id}/commit:
    post:
      description: Requires import:approve. Blocked (422) while any row is unresolved
        (balance migrations must also reconcile to the source total). Applies
        all rows in ONE transaction; re-commit is a no-op.
      operationId: ImportController_commit_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      security: *a45
      summary: Commit the batch (atomic + idempotent; reconcile-gated)
      tags: *a46
...
info:
  title: Redwave ERP / HRM API
  description: Versioned REST contract for the Redwave platform. Each protected
    endpoint notes the required (module, action) permission enforced server-side
    by the RBAC guard.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateRepDto:
      type: object
      properties:
        rep_code:
          type: string
          example: Redwave07
          description: Unique rep code. Never reused — even after termination.
        full_name:
          type: string
          example: Jordan Field
        field_manager_id:
          type: string
          description: User id of the Field Manager (must hold the Manager role).
        hire_date:
          type: string
          example: 2026-01-15
          description: YYYY-MM-DD.
        user_id:
          type: string
          description: Optional linked login user (source of contact info).
        payment_details:
          type: object
          description: Sensitive banking / e-transfer details (role-gated).
      required:
        - rep_code
        - full_name
        - field_manager_id
        - hire_date
    UpdateRepDto:
      type: object
      properties:
        full_name:
          type: string
        field_manager_id:
          type: string
          description: Reassign the Field Manager (must hold the Manager role).
        hire_date:
          type: string
          example: 2026-01-15
        user_id:
          type: string
        payment_details:
          type: object
        status:
          type: string
          enum:
            - active
            - terminated
          description: terminated requires termination_date.
        termination_date:
          type: string
          example: 2026-06-30
          description: Required when status=terminated.
    CreateRepDocumentDto:
      type: object
      properties:
        doc_type:
          type: string
          example: contract
          description: e.g. contract, id, equipment_agreement.
        file_url:
          type: string
          description: Object-storage reference (URL/key) of the stored file.
      required:
        - doc_type
        - file_url
    CreateRepEquipmentDto:
      type: object
      properties:
        equipment_type:
          type: string
          example: iPad
        identifier:
          type: string
          example: SN-12345
          description: Serial / asset tag.
        deposit_amount:
          type: string
          example: "250.00"
          description: Deposit held — exact decimal STRING, never a float.
        assigned_date:
          type: string
          example: 2026-01-15
          description: YYYY-MM-DD.
      required:
        - equipment_type
        - identifier
        - deposit_amount
        - assigned_date
...
    CreateImportDto:
      type: object
      properties:
        source_type:
          type: string
          enum:
            - client_report
            - master_migration
            - balance_migration
          example: client_report
        import_type:
          type: string
          enum:
            - reps
            - clients
            - products
            - sales
            - holdback
            - clawback
            - mixed
          example: sales
        client_id:
          type: string
          description: Client scope (required for client_report).
        field_mapping_id:
          type: string
          description: A saved field mapping to apply to each row.
        reconcile_total:
          type: string
          example: "48200.00"
          description: Operator-provided source total (required to commit a balance
            migration). Decimal string.
        rows:
          description: Raw source rows (file upload stubbed). Each is an arbitrary
            key→value object.
          type: array
          items:
            type: object
      required:
        - source_type
        - import_type
        - rows
    RowResolution:
      type: object
      properties:
        row_id:
          type: string
          description: The import row to resolve.
        action:
          type: string
          enum:
            - match
            - edit
            - ignore
        matched_entity_id:
          type: string
          description: "For 'match': the live entity to match this row to."
        mapped_data:
          type: object
          description: "For 'edit': corrected mapped data (re-classified)."
      required:
        - row_id
        - action
    ReconcileDto:
      type: object
      properties:
        resolutions:
          type: array
          items:
            $ref: "#/components/schemas/RowResolution"
      required:
        - resolutions
...